begin(green): command parser spikes
