+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+  "typescript.preferences.preferTypeOnlyAutoImports": true,
+  "typescript.format.semicolons": "insert"
+}
